Made separate .py files for each class, cleaned up the package
